+++ catalog/service/impl/src/test/java/org/projectnessie/catalog/service/impl/TestCatalogServiceImpl.java
@@ -15,16 +15,47 @@
  */
 package org.projectnessie.catalog.service.impl;
 
+import static java.util.Arrays.asList;
+import static java.util.concurrent.TimeUnit.MINUTES;
+import static org.assertj.core.api.InstanceOfAssertFactories.STRING;
 import static org.projectnessie.api.v2.params.ParsedReference.parsedReference;
+import static org.projectnessie.catalog.service.api.SnapshotReqParams.forSnapshotHttpReq;
+import static org.projectnessie.model.Content.Type.ICEBERG_TABLE;
+import static org.projectnessie.services.authz.Check.CheckType.COMMIT_CHANGE_AGAINST_REFERENCE;
+import static org.projectnessie.services.authz.Check.CheckType.READ_ENTITY_VALUE;
+import static org.projectnessie.services.authz.Check.CheckType.UPDATE_ENTITY;
+import static org.projectnessie.services.authz.Check.CheckType.VIEW_REFERENCE;
 
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.concurrent.atomic.AtomicReference;
+import java.util.function.Function;
+import java.util.stream.Collectors;
+import org.assertj.core.api.AbstractThrowableAssert;
 import org.junit.jupiter.api.Test;
 import org.projectnessie.api.v2.params.ParsedReference;
+import org.projectnessie.catalog.formats.iceberg.meta.IcebergJson;
+import org.projectnessie.catalog.formats.iceberg.meta.IcebergTableMetadata;
+import org.projectnessie.catalog.formats.iceberg.nessie.NessieModelIceberg;
+import org.projectnessie.catalog.model.snapshot.NessieTableSnapshot;
 import org.projectnessie.catalog.service.api.CatalogCommit;
+import org.projectnessie.catalog.service.api.CatalogService;
+import org.projectnessie.catalog.service.api.SnapshotReqParams;
+import org.projectnessie.catalog.service.api.SnapshotResponse;
+import org.projectnessie.model.Content;
 import org.projectnessie.model.ContentKey;
+import org.projectnessie.model.IcebergTable;
 import org.projectnessie.model.Reference;
+import org.projectnessie.services.authz.AbstractBatchAccessChecker;
+import org.projectnessie.services.authz.AccessCheckException;
+import org.projectnessie.services.authz.Check;
+import org.projectnessie.services.authz.Check.CheckType;
+import org.projectnessie.storage.uri.StorageUri;
 
 public class TestCatalogServiceImpl extends AbstractCatalogService {
 
+  /** Check that a non-modifying catalog-commit is a no-op. */
   @Test
   public void noCommitOps() throws Exception {
     Reference main = api.getReference().refName("main").get();
@@ -39,6 +70,7 @@ public class TestCatalogServiceImpl extends AbstractCatalogService {
     soft.assertThat(afterCommit).isEqualTo(main);
   }
 
+  /** Verifies that a single table create catalog-commit passes. */
   @Test
   public void singleTableCreate() throws Exception {
     Reference main = api.getReference().refName("main").get();
@@ -51,5 +83,123 @@ public class TestCatalogServiceImpl extends AbstractCatalogService {
         .isNotEqualTo(main)
         .extracting(Reference::getName, Reference::getHash)
         .containsExactly(committed.name(), committed.hashWithRelativeSpec());
+
+    SnapshotResponse snap =
+        catalogService
+            .retrieveSnapshot(
+                forSnapshotHttpReq(committed, "ICEBERG", "2"), key, ICEBERG_TABLE, false)
+            .toCompletableFuture()
+            .get(5, MINUTES);
+
+    soft.assertThat(snap)
+        .extracting(
+            SnapshotResponse::contentKey,
+            SnapshotResponse::contentType,
+            SnapshotResponse::effectiveReference)
+        .containsExactly(key, "application/json", afterCommit);
+
+    soft.assertThat(snap.content())
+        .extracting(IcebergTable.class::cast)
+        .extracting(IcebergTable::getMetadataLocation, STRING)
+        .endsWith(".metadata.json");
+    soft.assertThat(snap.entityObject()).containsInstanceOf(IcebergTableMetadata.class);
+
+    IcebergTableMetadata icebergMetadataEntity =
+        (IcebergTableMetadata) snap.entityObject().orElseThrow();
+
+    IcebergTableMetadata icebergMetadata =
+        NessieModelIceberg.nessieTableSnapshotToIceberg(
+            (NessieTableSnapshot) snap.nessieSnapshot(),
+            Optional.empty(),
+            m -> m.putAll(icebergMetadataEntity.properties()));
+
+    soft.assertThat(snap.entityObject()).contains(icebergMetadata);
+
+    String expectedJson =
+        IcebergJson.objectMapper()
+            .writeValueAsString(
+                IcebergTableMetadata.builder()
+                    .from(icebergMetadataEntity)
+                    .properties(Map.of())
+                    .build());
+    soft.assertThat(
+            objectIO.readObject(
+                StorageUri.of(((IcebergTable) snap.content()).getMetadataLocation())))
+        .hasContent(expectedJson);
+  }
+
+  /**
+   * Verify behavior of {@link CatalogService#retrieveSnapshot(SnapshotReqParams, ContentKey,
+   * Content.Type, boolean)} against related Nessie {@link CheckType check types} for read and write
+   * intents.
+   */
+  @Test
+  public void retrieveSnapshotAccessChecks() throws Exception {
+    Reference main = api.getReference().refName("main").get();
+    ContentKey key = ContentKey.of("mytable");
+
+    ParsedReference committed = commitSingle(main, key);
+
+    AtomicReference<CheckType> failingCheckType = new AtomicReference<>();
+    batchAccessCheckerFactory =
+        x ->
+            new AbstractBatchAccessChecker() {
+              @Override
+              public Map<Check, String> check() {
+                return getChecks().stream()
+                    .filter(c -> failingCheckType.get() == c.type())
+                    .collect(Collectors.toMap(Function.identity(), Object::toString));
+              }
+            };
+
+    List<CheckType> checks =
+        asList(
+            VIEW_REFERENCE,
+            COMMIT_CHANGE_AGAINST_REFERENCE,
+            READ_ENTITY_VALUE,
+            UPDATE_ENTITY,
+            null);
+    for (CheckType checkType : checks) {
+      boolean readFail = checkType == VIEW_REFERENCE || checkType == READ_ENTITY_VALUE;
+      boolean writeFail =
+          readFail || checkType == COMMIT_CHANGE_AGAINST_REFERENCE || checkType == UPDATE_ENTITY;
+      failingCheckType.set(checkType);
+
+      AbstractThrowableAssert<?, ? extends Throwable> read =
+          soft.assertThatCode(
+                  () ->
+                      catalogService
+                          .retrieveSnapshot(
+                              forSnapshotHttpReq(committed, "ICEBERG", "2"),
+                              key,
+                              ICEBERG_TABLE,
+                              false)
+                          .toCompletableFuture()
+                          .get(5, MINUTES))
+              .describedAs("forRead with %s", checkType);
+      if (readFail) {
+        read.isInstanceOf(AccessCheckException.class);
+      } else {
+        read.doesNotThrowAnyException();
+      }
+
+      AbstractThrowableAssert<?, ? extends Throwable> write =
+          soft.assertThatCode(
+                  () ->
+                      catalogService
+                          .retrieveSnapshot(
+                              forSnapshotHttpReq(committed, "ICEBERG", "2"),
+                              key,
+                              ICEBERG_TABLE,
+                              true)
+                          .toCompletableFuture()
+                          .get(5, MINUTES))
+              .describedAs("forWrite with %s", checkType);
+      if (writeFail) {
+        write.isInstanceOf(AccessCheckException.class);
+      } else {
+        write.doesNotThrowAnyException();
+      }
+    }
   }
 }