+++ catalog/service/rest/src/test/java/org/projectnessie/catalog/service/rest/TestIcebergS3SignParams.java
@@ -119,7 +119,8 @@ class TestIcebergS3SignParams {
   @ParameterizedTest
   @ValueSource(strings = {"GET", "HEAD", "OPTIONS", "TRACE"})
   void verifyAndSignSuccessRead(String method) throws Exception {
-    when(catalogService.retrieveSnapshot(any(), eq(key), isNull())).thenReturn(successStage);
+    when(catalogService.retrieveSnapshot(any(), eq(key), isNull(), eq(false)))
+        .thenReturn(successStage);
     when(signer.sign(any())).thenReturn(signingResponse);
     IcebergS3SignParams icebergSigner =
         newBuilder()
@@ -132,7 +133,8 @@ class TestIcebergS3SignParams {
   @ParameterizedTest
   @ValueSource(strings = {"PUT", "POST", "DELETE", "PATCH"})
   void verifyAndSignSuccessWrite(String method) throws Exception {
-    when(catalogService.retrieveSnapshot(any(), eq(key), isNull())).thenReturn(successStage);
+    when(catalogService.retrieveSnapshot(any(), eq(key), isNull(), eq(true)))
+        .thenReturn(successStage);
     when(signer.sign(any())).thenReturn(signingResponse);
     IcebergS3SignParams icebergSigner =
         newBuilder()
@@ -164,7 +166,7 @@ class TestIcebergS3SignParams {
             key,
             view,
             nessieViewSnapshot);
-    when(catalogService.retrieveSnapshot(any(), eq(key), isNull()))
+    when(catalogService.retrieveSnapshot(any(), eq(key), isNull(), eq(true)))
         .thenReturn(CompletableFuture.completedStage(snapshotResponse));
     when(signer.sign(any())).thenReturn(signingResponse);
     IcebergS3SignParams icebergSigner = newBuilder().build();
@@ -174,7 +176,7 @@ class TestIcebergS3SignParams {
 
   @Test
   void verifyAndSignSuccessContentNotFound() throws Exception {
-    when(catalogService.retrieveSnapshot(any(), eq(key), isNull()))
+    when(catalogService.retrieveSnapshot(any(), eq(key), isNull(), eq(true)))
         .thenThrow(new NessieContentNotFoundException(key, "main"));
     when(signer.sign(any())).thenReturn(signingResponse);
     IcebergS3SignParams icebergSigner = newBuilder().build();
@@ -184,7 +186,7 @@ class TestIcebergS3SignParams {
 
   @Test
   void verifyAndSignFailureReferenceNotFound() throws Exception {
-    when(catalogService.retrieveSnapshot(any(), eq(key), isNull()))
+    when(catalogService.retrieveSnapshot(any(), eq(key), isNull(), eq(true)))
         .thenThrow(new NessieReferenceNotFoundException("ref not found"));
     IcebergS3SignParams icebergSigner = newBuilder().build();
     Uni<IcebergS3SignResponse> response = icebergSigner.verifyAndSign();
@@ -195,7 +197,8 @@ class TestIcebergS3SignParams {
   void verifyAndSignSuccessImportFailed() throws Exception {
     CompletionStage<SnapshotResponse> importFailedStage =
         CompletableFuture.failedStage(new RuntimeException("import failed"));
-    when(catalogService.retrieveSnapshot(any(), eq(key), isNull())).thenReturn(importFailedStage);
+    when(catalogService.retrieveSnapshot(any(), eq(key), isNull(), eq(true)))
+        .thenReturn(importFailedStage);
     when(signer.sign(any())).thenReturn(signingResponse);
     IcebergS3SignParams icebergSigner = newBuilder().build();
     Uni<IcebergS3SignResponse> response = icebergSigner.verifyAndSign();
@@ -205,7 +208,8 @@ class TestIcebergS3SignParams {
   @ParameterizedTest
   @ValueSource(strings = {"GET", "HEAD", "OPTIONS", "TRACE"})
   void verifyAndSignSuccessReadMetadataLocation(String method) throws Exception {
-    when(catalogService.retrieveSnapshot(any(), eq(key), isNull())).thenReturn(successStage);
+    when(catalogService.retrieveSnapshot(any(), eq(key), isNull(), eq(false)))
+        .thenReturn(successStage);
     when(signer.sign(any())).thenReturn(signingResponse);
     IcebergS3SignParams icebergSigner =
         newBuilder()
@@ -218,7 +222,8 @@ class TestIcebergS3SignParams {
   @ParameterizedTest
   @ValueSource(strings = {"PUT", "POST", "DELETE", "PATCH"})
   void verifyAndSignFailureWriteMetadataLocation(String method) throws Exception {
-    when(catalogService.retrieveSnapshot(any(), eq(key), isNull())).thenReturn(successStage);
+    when(catalogService.retrieveSnapshot(any(), eq(key), isNull(), eq(true)))
+        .thenReturn(successStage);
     IcebergS3SignParams icebergSigner =
         newBuilder()
             .request(
@@ -235,7 +240,8 @@ class TestIcebergS3SignParams {
   @ParameterizedTest
   @ValueSource(strings = {"GET", "HEAD", "OPTIONS", "TRACE"})
   void verifyAndSignSuccessReadAncientLocation(String method) throws Exception {
-    when(catalogService.retrieveSnapshot(any(), eq(key), isNull())).thenReturn(successStage);
+    when(catalogService.retrieveSnapshot(any(), eq(key), isNull(), eq(false)))
+        .thenReturn(successStage);
     when(signer.sign(any())).thenReturn(signingResponse);
     IcebergS3SignParams icebergSigner =
         newBuilder()
@@ -253,7 +259,8 @@ class TestIcebergS3SignParams {
   @ParameterizedTest
   @ValueSource(strings = {"PUT", "POST", "DELETE", "PATCH"})
   void verifyAndSignFailureWriteAncientLocation(String method) throws Exception {
-    when(catalogService.retrieveSnapshot(any(), eq(key), isNull())).thenReturn(successStage);
+    when(catalogService.retrieveSnapshot(any(), eq(key), isNull(), eq(true)))
+        .thenReturn(successStage);
     IcebergS3SignParams icebergSigner =
         newBuilder()
             .request(
@@ -269,7 +276,8 @@ class TestIcebergS3SignParams {
 
   @Test
   void verifyAndSignFailureWrongBaseLocation() throws Exception {
-    when(catalogService.retrieveSnapshot(any(), eq(key), isNull())).thenReturn(successStage);
+    when(catalogService.retrieveSnapshot(any(), eq(key), isNull(), eq(true)))
+        .thenReturn(successStage);
     IcebergS3SignParams icebergSigner =
         newBuilder().baseLocation("s3://wrong-bucket/warehouse/ns/table1_cafebabee").build();
     Uni<IcebergS3SignResponse> response = icebergSigner.verifyAndSign();