+++ CHANGELOG.md
@@ -16,6 +16,7 @@ as necessary. Empty sections will not end in the release notes.
 
 - Catalog/GCS: Support using the default application credentials
 - Catalog/S3: Allow custom key+trust stores
+- Catalog: Check privileges earlier
 
 ### Changes